chore: split style in multiple files
