+++ apps/frontend/src/components/ActionCard/styled.tsx
@@ -1,0 +1,27 @@
+import styled from 'styled-components';
+
+export const Card = styled.div`
+  background-color: #fff;
+  border-radius: 8px;
+  box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
+  margin: 10px;
+  padding: 20px;
+  width: 300px;
+  text-align: center;
+`;
+
+export const Button = styled.button`
+  background-color: #6200ee;
+  border: none;
+  border-radius: 4px;
+  color: white;
+  cursor: pointer;
+  font-size: 16px;
+  margin-top: 10px;
+  padding: 10px 20px;
+`;
+
+export const Credit = styled.p`
+  font-size: 14px;
+  color: #666;
+`;
