+++ apps/frontend/src/pages/Home/index.tsx
@@ -1,15 +1,8 @@
 import React, { useEffect, useState } from 'react';
-import styled from 'styled-components';
-import ActionCard from '../components/ActionCard';
-import Navbar from '../components/Navbar';
-import { getActionTypes } from '../services/actionTypes';
-
-const Container = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  padding: 20px;
-`;
+import ActionCard from '../../components/ActionCard';
+import Navbar from '../../components/Navbar';
+import { getActionTypes } from '../../services/actionTypes';
+import { Container } from './styled';
 
 export default function HomePage() {
   const [actionTypes, setActionTypes] = useState([]);
