+++ apps/frontend/src/pages/Home/styled.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+  padding: 20px;
+`;
